requirements.txt
streamlit
openai
langchain
chromadb
PyPDF2
langchain-chroma
streamlit-chat
python-dotenv